# priv/static/swagger.json
{
  "definitions": {
...
    "GroupSessions": {
      "description": "Relation between group and session",
      "example": {
        "group_id": 1,
        "session_id": 1
      },
      "properties": {
        "group_id": {
          "description": "Id of a particular group",
          "type": "integer"
        },
        "session_id": {
          "description": "Id of a particular group",
          "type": "integer"
        }
      },
      "title": "Group Session",
      "type": "object"
    },
...
    "SessionIds": {
      "example": {
        "session_ids": [
          1,
          2
        ]
      },
      "properties": {
        "session_ids": {
          "description": "List of session ids",
          "type": "array"
        }
      },
      "type": "object"
    },
...
    "GroupType": {
      "description": "A Group type in application",
      "example": {
        "child_id": {
          "donor": "YES",
          "group_id": 29,
          "id": 24,
          "mode": "Offline",
          "model": "Live Classes",
          "name": "Mrs. Stefanie Goldner",
          "product_used": "One",
          "start_date": "2016-11-18",
          "state": "UTTARAKHAND",
          "sub_type": "High",
          "target_outreach": 4743,
          "type": "Competitive"
        },
        "id": 54,
        "type": "program"
      },
      "properties": {
        "child_id": {
          "description": "The id of type",
          "type": "integer"
        },
        "type": {
          "description": "The type of a group",
          "type": "string"
        }
      },
      "title": "GroupType",
      "type": "object"
    },
...
    "GroupUsers": {
      "description": "Relation between group and user",
      "example": {
        "group_id": 1,
        "program_date_of_joining": "2020/01/06",
        "program_manager_id": 1,
        "program_student_language": "English",
        "user_id": 1
      },
      "properties": {
        "group_id": {
          "description": "Id of a particular group",
          "type": "integer"
        },
        "program_date_of_joining": {
          "description": "Date of joining a program",
          "type": "utc_datetime"
        },
        "program_manager_id": {
          "description": "ID of a program manager",
          "type": "integer"
        },
        "program_student_language": {
          "description": "Language used in an enrolled program",
          "type": "string"
        },
        "user_id": {
          "description": "Id of a particular group",
          "type": "integer"
        }
      },
      "title": "Group User",
      "type": "object"
    },
...
    "GroupTypes": {
      "description": "All the GroupTypes",
      "items": {
        "$ref": "#/definitions/GroupType"
      },
      "title": "GroupTypes",
      "type": "array"
    },
...
    "UserIds": {
      "example": {
        "user_ids": [
          1,
          2
        ]
      },
      "properties": {
        "user_ids": {
          "description": "List of user ids",
          "type": "array"
        }
      },
      "type": "object"
    },
...
    "Group": {
      "description": "A group in the application",
      "example": {
        "input_schema": {},
        "locale": "hi",
        "locale_data": {
          "en": {
            "title": "Register for session"
          },
          "hi": {
            "title": "सत्र के लिए पंजीकरण करें"
          }
        },
        "name": "DelhiStudents"
      },
      "properties": {
        "input_schema": {
          "description": "Input schema",
          "type": "map"
        },
        "locale": {
          "description": "The configured locale for the group",
          "type": "string"
        },
        "locale_data": {
          "description": "Meta data about locale settings for the group",
          "type": "map"
        },
        "name": {
          "description": "Name of a group",
          "type": "string"
        }
      },
      "title": "Group",
      "type": "object"
    },
...
  },
  "info": {
    "title": "DB Service application",
    "version": "1.0"
  },
  "paths": {
...
    "/api/group-session/{groupSessionId}": {
      "delete": {
        "description": "",
        "operationId": "DbserviceWeb.GroupSessionController.delete",
        "parameters": [
          {
            "description": "The id of the group session record",
            "in": "path",
            "name": "groupSessionId",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "summary": "",
        "tags": [
          "GroupSession"
        ]
      },
      "get": {
        "description": "",
        "operationId": "DbserviceWeb.GroupSessionController.show",
        "parameters": [
          {
            "description": "The id of the group session record",
            "in": "path",
            "name": "groupSessionId",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/GroupSessions"
            }
          }
        },
        "summary": "",
        "tags": [
          "GroupSession"
        ]
      },
      "patch": {
        "description": "",
        "operationId": "DbserviceWeb.GroupSessionController.update",
        "parameters": [
          {
            "description": "The id of the session record",
            "in": "path",
            "name": "groupSessionId",
            "required": true,
            "type": "integer"
          },
          {
            "description": "Group session to create",
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/GroupSessions"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated",
            "schema": {
              "$ref": "#/definitions/GroupSessions"
            }
          }
        },
        "summary": "",
        "tags": [
          "GroupSession"
        ]
      }
    },
...
    "/api/group-type/{groupTypeId}/update-users": {
      "post": {
        "description": "",
        "operationId": "DbserviceWeb.GroupTypeController.update_users",
        "parameters": [
          {
            "description": "The id of the group record",
            "in": "path",
            "name": "groupTypeId",
            "required": true,
            "type": "integer"
          },
          {
            "description": "List of user ids to update for the group",
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserIds"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/GroupUsers"
            }
          }
        },
        "summary": "",
        "tags": [
          "GroupType"
        ]
      }
    },
...
    "/api/group-type/{groupTypeId}": {
      "delete": {
        "description": "",
        "operationId": "DbserviceWeb.GroupTypeController.delete",
        "parameters": [
          {
            "description": "The id of the group-type record",
            "in": "path",
            "name": "groupTypeId",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "summary": "",
        "tags": [
          "GroupType"
        ]
      },
      "get": {
        "description": "",
        "operationId": "DbserviceWeb.GroupTypeController.show",
        "parameters": [
          {
            "description": "The id of the group-type record",
            "in": "path",
            "name": "groupTypeId",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/GroupType"
            }
          }
        },
        "summary": "",
        "tags": [
          "GroupType"
        ]
      },
      "patch": {
        "description": "",
        "operationId": "DbserviceWeb.GroupTypeController.update",
        "parameters": [
          {
            "description": "The id of the group-type record",
            "in": "path",
            "name": "groupTypeId",
            "required": true,
            "type": "integer"
          },
          {
            "description": "Group to create",
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/GroupType"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated",
            "schema": {
              "$ref": "#/definitions/Group"
            }
          }
        },
        "summary": "",
        "tags": [
          "GroupType"
        ]
      }
    },
...
    "/api/group-type": {
      "get": {
        "description": "",
        "operationId": "DbserviceWeb.GroupTypeController.index",
        "parameters": [
          {
            "description": "The type the group",
            "in": "query",
            "name": "type",
            "required": false,
            "type": "string"
          },
          {
            "description": "The child id the group type",
            "in": "query",
            "name": "child_id",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/GroupTypes"
            }
          }
        },
        "summary": "",
        "tags": [
          "GroupType"
        ]
      },
      "post": {
        "description": "",
        "operationId": "DbserviceWeb.GroupTypeController.create",
        "parameters": [
          {
            "description": "Group to create",
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/GroupType"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/GroupType"
            }
          }
        },
        "summary": "",
        "tags": [
          "GroupType"
        ]
      }
    },
    "/api/group-type/{groupTypeId}/update-sessions": {
      "post": {
        "description": "",
        "operationId": "DbserviceWeb.GroupTypeController.update_sessions",
        "parameters": [
          {
            "description": "The id of the group record",
            "in": "path",
            "name": "groupTypeId",
            "required": true,
            "type": "integer"
          },
          {
            "description": "List of session ids to update for the group",
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SessionIds"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/GroupSessions"
            }
          }
        },
        "summary": "",
        "tags": [
          "GroupType"
        ]
      }
    },
...
    "/api/group-session": {
      "get": {
        "description": "",
        "operationId": "DbserviceWeb.GroupSessionController.index",
        "parameters": [
          {
            "description": "The id the group_type",
            "in": "query",
            "name": "group_type_id",
            "required": false,
            "type": "string"
          },
          {
            "description": "The id the session",
            "in": "query",
            "name": "session_id",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/GroupSessions"
            }
          }
        },
        "summary": "",
        "tags": [
          "GroupSession"
        ]
      },
      "post": {
        "description": "",
        "operationId": "DbserviceWeb.GroupSessionController.create",
        "parameters": [
          {
            "description": "Group session to create",
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/GroupSessions"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/GroupSessions"
            }
          }
        },
        "summary": "",
        "tags": [
          "GroupSession"
        ]
      }
    }
  },
  "swagger": "2.0"
}